+++ CHANGELOG.md
@@ -32,6 +32,9 @@ Converting from v3.x to v4.0
   also use the `extend` variants. Otherwise, you will enter into an infinite recursion.
 - OpenSim now makes substantial use of C++11 features; if you compile OpenSim, your compiler
   must support C++11. Also, any C++ project in which you use OpenSim must also be compiled with C++11.
+- The following components have been upgraded to use Connectors to connect to
+  other components they depend on (instead of string properties):
+  - ContactGeometry (ContactSphere, ContactHalfSpace, ContactMesh)
 - Many of the methods in ScaleTool have now been marked const.
 
 Bug Fixes